Arquivos novos na nuvem
Upando para a nuvem todas as alterações já com o banco da escola

+++ routes/web.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\ControllerProfessores;
 use App\Http\Controllers\ControllerEscola;
 use Illuminate\Console\Scheduling\Event;
 use App\Http\Controllers\Controllerlogin;
+use App\Models\User;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -31,17 +32,21 @@ Route::get('/principal', [ControllerSisco::class, 'index'])->middleware('auth');
 
 Route::get('/turmas', [ControllerTurmas::class, 'index'])->middleware('auth');
 
-Route::get('/professores', [ControllerProfessores::class, 'index']);
+Route::get('/professores', [ControllerProfessores::class, 'index'])->middleware('auth');
 
 Route::get('/cadcoordenadores', function() {
-    return view('content.cadastro-coordenadores');
-});
+    $idUser = auth()->id();
+    $tipoUser = User::findOrFail($idUser)->tipo_user;
+    return view('content.cadastro-coordenadores', ['tipoUser' => $tipoUser]);
+})->middleware('auth');
 
-Route::get('/escola', [ControllerEscola::class, 'show']);
+Route::get('/escola', [ControllerEscola::class, 'show'])->middleware('auth');
 
 Route::get('/manual', function() {
-    return view('content.manual');
-});
+    $idUser = auth()->id();
+    $tipoUser = User::findOrFail($idUser)->tipo_user;
+    return view('content.manual',['tipoUser' => $tipoUser]);
+})->middleware('auth');
 
 Route::post('/ocorrencia', [ControllerSisco::class, 'store']);
 
@@ -51,12 +56,9 @@ Route::post('/editEscola', [ControllerEscola::class, 'edit']);
 
 Route::post('/novocoordenador', [ControllerCoordenadores::class, 'store']);
 
-// Estabelecer Route para o \/cadTurmas
-Route::get('/cadTurmas', [ControllerTurmas::class, 'register']);
-
 Route::post('/novaturma', [ControllerTurmas::class, 'store']);
 
-Route::get('/cadCurso', [ControllerCursos::class, 'index']);
+Route::get('/cadCurso', [ControllerCursos::class, 'index'])->middleware('auth');
 
 Route::post('/novocurso', [ControllerCursos::class, 'store']);
 
@@ -64,7 +66,7 @@ Route::get('/editProf/{id}', [ControllerProfessores::class, 'edit']);
 
 Route::delete('/delProf/{id}', [ControllerProfessores::class, 'destroy']);
 
-Route::get('/cadAlunos', [ControllerAluno::class, 'index']);
+Route::get('/cadAlunos', [ControllerAluno::class, 'index'])->middleware('auth');
 
 Route::post('/aluno', [ControllerAluno::class, 'store']);
 
@@ -72,11 +74,13 @@ Route::post('/aluno-csv', [ControllerAluno::class, 'storeCSV']);
 
 Route::post('/updateUser/{id}', [ControllerProfessores::class, 'update']);
 
-Route::get('/relturmas/{id}', [ControllerTurmas::class, 'relatorioIndex']);
+Route::get('/relturmas/{id}', [ControllerTurmas::class, 'relatorioIndex'])->middleware('auth');
 
 Route::get('/cadastro-por-csv', function(){
-    return view('content.cadastro-csv');
-});
+    $idUser = auth()->id();
+    $tipoUser = User::findOrFail($idUser)->tipo_user;
+    return view('content.cadastro-csv', ['tipoUser' => $tipoUser]);
+})->middleware('auth');
 
 Route::post('/autenticacao', [Controllerlogin::class, 'index']);
 
@@ -90,8 +94,24 @@ Route::middleware([
     })->name('dashboard');
 });
 
-Route::get('/consulta/{id}', [ControllerTurmas::class, 'search']);
+Route::get('/consulta/{id}', [ControllerTurmas::class, 'search'])->middleware('auth');
 
-Route::post('/concluido/{id}', [ControllerSisco::class, 'marked']);
+Route::post('/concluido/{id}', [ControllerSisco::class, 'marked'])->middleware('auth');
 
-Route::delete('/delAluno/{id}', [ControllerAluno::class, 'destroy']);
+Route::delete('/delAluno/{id}', [ControllerAluno::class, 'destroy'])->middleware('auth');
+
+Route::get('/editAluno/{id}', [ControllerAluno::class, 'show'])->middleware('auth');
+
+Route::put('/updateAluno/{id}', [ControllerAluno::class, 'update'])->middleware('auth');
+
+Route::post('/logOut', [ControllerSisco::class, 'logout'])->middleware('auth');
+
+Route::get('/forgotPassword', function() {
+    return view('auth.email');
+});
+
+Route::post('/confirmLogin', [ControllerSisco::class, 'confirmLogin']);
+
+Route::post('/confirmNumber', [ControllerSisco::class, 'confirmNumber']);
+
+Route::post('/newPassword', [ControllerSisco::class, 'newPass']);
